backend/requirements.txt
# --- Web Framework ---
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9

# --- Deep Learning Engine ---
# Note: This is the CPU version. For GPU, use paddlepaddle-gpu
paddlepaddle==2.6.2
pycocotools==2.0.6
Levenshtein==0.12.2

# --- OCR & Computer Vision ---
paddleocr==2.7.3
opencv-python-headless==4.9.0.80
numpy==1.26.4
Pillow==10.2.0

# --- Data Synthesis (For Papyr Training) ---

#Manually install trdg due to dependency issues
#python -m pip install trdg==1.6.0 --no-deps 

#Manually install the OTHER missing dependencies trdg needs (excluding the problematic Pillow)
#python -m pip install opencv-python-headless numpy requests tqdm diffimg arabic-reshaper python-bidi wikipedia

imgaug==0.4.0